requirements.txt
aniso8601==1.2.1
asn1crypto==0.22.0
backports-abc==0.5
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
coverage==4.4.1
cryptography==2.0.3
enum34==1.1.6
Flask==0.12.2
Flask-RESTful==0.3.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
gtextfsm==0.2.1
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
junos-eznc==2.1.5
lxml==3.8.0
MarkupSafe==1.0
mock==2.0.0
ncclient==0.5.3
netaddr==0.7.19
netmiko==1.4.2
paramiko==2.2.1
pbr==3.1.1
progress==1.3
pyasn1==0.3.3
pycparser==2.18
pyeapi==0.8.1
PyNaCl==1.1.2
pyntc==0.0.5
pynxos==0.0.3
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scp==0.10.2
singledispatch==3.4.0.3
six==1.10.0
sockjs-tornado==1.0.3
terminal==0.4.0
urllib3==1.22
Werkzeug==0.12.2
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
psycopg2
celery
redis